requirements-production.txt
Flask==2.3.3
PyMySQL==1.1.0
python-dotenv==1.0.0

# Production WSGI Server
gunicorn==21.2.0

# Security
bcrypt==4.0.1

# Optional: Redis for session storage
# redis==5.0.1

# Optional: Email support
# Flask-Mail==0.9.1

# Optional: Rate limiting
# Flask-Limiter==3.5.0

# Optional: HTTPS redirect
# flask-talisman==1.1.0
